Aldatu egin da float double motara
Irabaziak metodoak bueltatzen zuen float bat, baina arazoa hematen zuen, orduan double motara aldatu egin da.

+++ UD5 - OBP aurreratua/1. Ariketa multzoa/_3_ariketa/src/KomisioLangile.java
@@ -16,7 +16,7 @@ public class KomisioLangile extends Langile{
         }
     }
 
-    public float irabaziak(){
+    public double irabaziak(){
         return this.salmentak*this.komisioa;
     }
