+++ UD5 - OBP aurreratua/1. Ariketa multzoa/_3_ariketa/src/SoldatapekoLangile.java
@@ -11,7 +11,7 @@ public class SoldatapekoLangile extends Langile{
 
 
     }
-    public float irabaziak(){
+    public double irabaziak(){
         return astekoSoldata;
     }
